Deleted transferArray, Update main and identicalChar
Checking the transferred elements in identicalChar, and a small adjustment of main

+++ kiselev.sergey/P4/identicalChar.h
@@ -3,6 +3,6 @@
 #include <cstddef>
 namespace kiselev
 {
-  int identicalChar(char* arr1, const char* arr2, size_t lenght1, size_t lenght2);
+  size_t identicalChar(char* arr1, const char* arr2, size_t lenght1, size_t lenght2);
 }
 #endif
